+++ src/components/Table/components/CellRenderer/components/GaugeCellRenderer/GaugeCellRenderer.tsx
@@ -19,7 +19,7 @@ interface Props {
   value: number;
 
   /**
-   * Bg Color
+   * Background Color
    *
    * @type {string}
    */